Refactor test helpers
