+++ tests/test_gl_codes.py
@@ -1,7 +1,7 @@
 from werkzeug.security import generate_password_hash
 from app import db
 from app.models import User, GLCode, Item, ItemUnit, Product, ProductRecipeItem
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def setup_data(app):
