+++ tests/test_import_items.py
@@ -3,7 +3,7 @@ from werkzeug.security import generate_password_hash
 
 from app import db
 from app.models import User
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def create_user(app, email='import@example.com'):
