+++ tests/test_invoice_filtering.py
@@ -3,7 +3,7 @@ from werkzeug.security import generate_password_hash
 
 from app import db
 from app.models import User, Customer, Invoice
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def setup_invoices(app):
