+++ tests/test_multiple_units.py
@@ -1,7 +1,7 @@
 from werkzeug.security import generate_password_hash
 from app import db
 from app.models import User, Item, ItemUnit
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def create_user(app, email='multi@example.com'):
