+++ tests/test_product_recipe.py
@@ -1,7 +1,7 @@
 from werkzeug.security import generate_password_hash
 from app import db
 from app.models import User, Customer, Item, ItemUnit, Product, ProductRecipeItem, Invoice
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def setup_data(app):
