+++ tests/test_purchase_flow.py
@@ -14,7 +14,7 @@ from app.models import (
     PurchaseOrderItemArchive,
     GLCode,
 )
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def setup_purchase(app):
