+++ tests/test_sales_invoice.py
@@ -4,7 +4,7 @@ from werkzeug.security import generate_password_hash
 
 from app import db
 from app.models import User, Customer, Product, Invoice
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def setup_sales(app):
