+++ tests/test_vendor_crud.py
@@ -2,7 +2,7 @@ from werkzeug.security import generate_password_hash
 
 from app import db
 from app.models import User, Vendor
-from tests.test_user_flows import login
+from tests.utils import login
 
 
 def setup_user(app):
